FEAT: Move files in common
