+++ trieTools/zeroBalanceSystemAccountChecker/interface.go
@@ -1,11 +1,5 @@
 package main
 
-import (
-	"io"
-
-	"github.com/ElrondNetwork/elrond-tools-go/trieTools/zeroBalanceSystemAccountChecker/common"
-)
-
 type crossTokenChecker interface {
 	crossCheckExtraTokens(tokens map[string]struct{}) ([]string, error)
 }
@@ -16,11 +10,4 @@ type tokenBalancesGetter interface {
 
 type elasticMultiSearchClient interface {
 	GetMultiple(index string, requests []string) ([]byte, error)
-}
-
-type fileHandler interface {
-	Open(name string) (io.Reader, error)
-	ReadAll(r io.Reader) ([]byte, error)
-	Getwd() (dir string, err error)
-	ReadDir(dirname string) ([]common.FileInfo, error)
 }
